New package structure and imported ProFit features
Refactored package to have python interfaces in the python directory, which will be changed later so the import paths are less silly. Also added lots of ProFit-like functionality and heavily modified the benchmarking script to use it while profiling integration and convolution simultaneously, as well as introducing a galsim backend option.